dt-bindings: power: rockchip: add regulator support
Add optional support for a voltage supply required to enable a
power domain. The binding follows the way it is handled by the
Mediatek binding to keep things consistent.

This will initially be used by the RK3588 GPU power domain, which
fails to be enabled when the GPU regulator is not enabled.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>